refactoring: extracted ITs with Tools in a separate test class
